some finals changes
